+++ src/extension/types.ts
@@ -17,6 +17,11 @@ export type JumpToDefinition = {
   col: number
 }
 
+export enum CvlVersion {
+  cvlVersion1 = 1,
+  cvlVersion2 = 2,
+}
+
 export enum RuleStatuses {
   Verified = 'VERIFIED',
   Violated = 'VIOLATED',
@@ -258,7 +263,7 @@ export type EventFromResultsWebview =
     }
   | {
       command: CommandFromResultsWebview.GetOutput
-      payload: string
+      payload: { url: string; name: string }
     }
   | {
       command: CommandFromResultsWebview.OpenLogFile
@@ -338,7 +343,7 @@ export type EventFromSettingsWebview =
 
 export type ConfFile = {
   files?: string[]
-  verify?: string
+  verify?: string | string[]
   solc?: string
   link?: string[]
   settings?: string[]